mis a jour des autorisations

+++ src/Controller/TrickController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 
 
+
 use App\Entity\Trick;
 use DateTimeInterface;
 use App\Entity\Message;
@@ -13,6 +14,7 @@ use App\Repository\UserRepository;
 use App\Repository\TrickRepository;
 use App\Repository\MessageRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -23,7 +25,10 @@ class TrickController extends AbstractController
     /**
      * @Route("/trick/{slug}", name="trick")
      */
-    public function showTrick($slug, Request $request, TrickRepository $trickRepository, Trick $trick, MessageRepository $messageRepository,  UserRepository $userRepository): Response
+    public function showTrick($slug, Request $request, TrickRepository $trickRepository, 
+        Trick $trick, MessageRepository $messageRepository,  
+        UserRepository $userRepository,
+        PaginatorInterface $paginator): Response
     {       
         $trick = $trickRepository->findOneBy(['slug' =>$slug]);
 
@@ -34,6 +39,12 @@ class TrickController extends AbstractController
         $message = new Message();
         $form = $this->createForm(MessageType::class, $message);
         $form->handleRequest($request);
+
+        $result = $paginator->paginate(
+            $trick->getMessage(),
+            $request->query->getInt('page', 1),
+            $request->query->getInt('limit', 10)
+        );
 
         //Traitement du formulaire         
         if ($form->isSubmitted() && $form->isValid()) {
@@ -49,7 +60,7 @@ class TrickController extends AbstractController
 
         return $this->render('trick/show.html.twig', [
             'trick' => $trick,
-            //'message' => $message,
+            'messages' => $result,
             'form'=> $form->createView()
         ]);
     }
@@ -88,25 +99,27 @@ class TrickController extends AbstractController
      */
     public function edit(Request $request, Trick $trick, TrickRepository $trickRepository): Response
     {
-        // $this->denyAccessUnlessGranted('trick_edit', $trick, 'Vous ne pouvez pas modifier cette figure');
-        $form = $this->createForm(TrickType::class, $trick);
-        $form->handleRequest($request);
+            $this->denyAccessUnlessGranted('trick_edit', $trick);       
+        
+            $form = $this->createForm(TrickType::class, $trick);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            //$trick->setUser($this->getUser());   
+            if ($form->isSubmitted() && $form->isValid()) {
+                //$trick->setUser($this->getUser());   
 
-            $trickRepository->add($trick, true);
+                $trickRepository->add($trick, true);
 
-            return $this->redirectToRoute('home', [], Response::HTTP_SEE_OTHER);
-        }
+                return $this->redirectToRoute('home', [], Response::HTTP_SEE_OTHER);
+            }
 
-        $flashMessage = $this->addFlash('success', 'Votre modification a ete enregistree !');
+            $flashMessage = $this->addFlash('success', 'Votre modification a ete enregistree !');
+
+        // $flashMessage = $this->addFlash('danger', 'Attention, vous ne pouvez pas modifier cette figure');
 
 
         return $this->renderForm('trick/edit.html.twig', [
             'trick' => $trick,
             'form' => $form
-            // 'flashMessage' => $flashMessage
         ]);
     }
 
@@ -116,7 +129,7 @@ class TrickController extends AbstractController
      */
     public function delete(Request $request, Trick $trick, TrickRepository $trickRepository): Response
     {
-        // $this->denyAccessUnlessGranted('trick_delete', $trick, 'Vous ne pouvez pas supprimer cette figure');
+        $this->denyAccessUnlessGranted('trick_delete', $trick, 'Vous ne pouvez pas supprimer cette figure');
 
         if ($this->isCsrfTokenValid('delete'.$trick->getId(), $request->request->get('_token'))) {
             $trickRepository->remove($trick, true);
